gastropartner-backend/pyproject.toml
[project]
name = "gastropartner"
version = "0.1.0"
description = "GastroPartner - SaaS för restauranger och livsmedelsproducenter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "email-validator>=2.2.0",
    "pyjwt>=2.10.1",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai>=0.8.0",
    "watchdog>=6.0.0",
    "rich>=14.1.0",
    "asyncio-throttle>=1.0.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = [
    "E501",  # Line length handled by formatter
    "B008",  # FastAPI dependency injection pattern is legitimate
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["src/gastropartner/tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
# Always generate reports, even on test failures
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    # Generate JUnit XML report
    "--junit-xml=reports/junit/pytest.xml",
    "--junit-prefix=gastropartner",
    # Generate coverage reports
    "--cov=src/gastropartner",
    "--cov-report=html:reports/htmlcov",
    "--cov-report=xml:reports/coverage.xml", 
    "--cov-report=json:reports/coverage.json",
    "--cov-report=term-missing",
    # Don't fail coverage on test failures
    "--no-cov-on-fail",
]
# Create reports directory structure
junit_family = "xunit2"
junit_logging = "all"
junit_log_passing_tests = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
    "pip-audit>=2.7.0",
    "pre-commit>=4.3.0",
]
